scripts/requirements.txt
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
Deprecated==1.2.12
idna==2.10
lxml==4.6.3
numpy==1.21.0
pandas==1.2.5
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.6
wrapt==1.12.1
